currency_dector/requirements.txt
tensorflow>=2.10.0
opencv-python>=4.5.0
matplotlib>=3.5.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
Pillow>=8.3.0
seaborn>=0.11.0
kagglehub
